Correct type casts

+++ bbr/bbr_linux.go
@@ -60,8 +60,8 @@ func getMaxBandwidthAndMinRTT(fp *os.File) (model.BBRInfo, error) {
 		return metrics, syscall.EOVERFLOW
 	}
 	metrics.BW = int64(maxbw) // Java has no uint64
-	metrics.MinRTT = int64(bbrip.bbr_min_rtt)
-	metrics.PacingGain = uint64(bbrip.bbr_pacing_gain)
-	metrics.CwndGain = uint64(bbrip.bbr_cwnd_gain)
+	metrics.MinRTT = uint32(bbrip.bbr_min_rtt)
+	metrics.PacingGain = uint32(bbrip.bbr_pacing_gain)
+	metrics.CwndGain = uint32(bbrip.bbr_cwnd_gain)
 	return metrics, nil
 }
